refactor(Calendar): seperate google login
