+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled, { ThemeProvider } from 'styled-components';
-import { GoogleOAuthProvider } from '@react-oauth/google';
+
 import globalTheme, { GlobalStyle } from './theme';
 import NavBar from './NavBar';
 import BackgroundImage from './BackgroundImage';
@@ -57,29 +57,21 @@ function App() {
 
   return (
     <ThemeProvider theme={globalTheme}>
-      <GoogleOAuthProvider
-        clientId={process.env.REACT_APP_GOOGLE_CLIENT_ID as string}
-      >
-        <GlobalStyle />
-        <BackgroundImage />
-        <Wrapper>
-          <NavBar addWidget={addWidget} availableWidgets={availableWidgets} />
-          <GridLayout
-            widgets={widgets}
-            layouts={layouts}
-            setLayouts={setLayouts}
-          >
-            {widgets.map((widget) => (
-              <Widget
-                key={widget.id}
-                widget={widget}
-                deleteWidget={() => deleteWidget(widget.id)}
-                onWidgetChange={onWidgetChange}
-              />
-            ))}
-          </GridLayout>
-        </Wrapper>
-      </GoogleOAuthProvider>
+      <GlobalStyle />
+      <BackgroundImage />
+      <Wrapper>
+        <NavBar addWidget={addWidget} availableWidgets={availableWidgets} />
+        <GridLayout widgets={widgets} layouts={layouts} setLayouts={setLayouts}>
+          {widgets.map((widget) => (
+            <Widget
+              key={widget.id}
+              widget={widget}
+              deleteWidget={() => deleteWidget(widget.id)}
+              onWidgetChange={onWidgetChange}
+            />
+          ))}
+        </GridLayout>
+      </Wrapper>
     </ThemeProvider>
   );
 }
