+++ src/hooks/useGoogleAccessToken/index.tsx
@@ -1,0 +1,3 @@
+import useGoogleAccessToken from './useGoogleAccessToken';
+
+export default useGoogleAccessToken;
